REFACTOR Tidy up message tests
The goal of these changes was to ensure that the tests for Message were
close to how it would want to be called (i.e. things should be specified
in the constructor rather than separately).  Also, the opportunity was
taken to do some general tidying around the Message code and tests.

* Make files shorter (by removing whitespace)
* Tidy up Message code
* Remove one Message test (was redundant/OTT)
* Similar tidying of whitespace in State tests.

Perhaps this was rather a lot for one commit :-).